entidad course api lista

+++ Backend/RH365.Core/Application/Features/DTOs/Course/CourseDto.cs
@@ -2,6 +2,8 @@
 // Archivo: CourseDto.cs
 // Proyecto: RH365.Core
 // Ruta: RH365.Core/Application/Features/DTOs/Course/CourseDto.cs
+// Descripción: DTO para transferencia de datos de Course.
+//   - Incluye todos los campos de la entidad más auditoría ISO 27001
 // ============================================================================
 using System;
 
@@ -14,7 +16,7 @@ namespace RH365.Core.Application.Features.DTOs.Course
         public string CourseCode { get; set; } = null!;
         public string Name { get; set; } = null!;
         public long CourseTypeRefRecID { get; set; }
-        public int? ClassRoomRefRecID { get; set; }
+        public long? ClassRoomRefRecID { get; set; }
         public string? Description { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
