+++ Backend/RH365.Core/Application/Features/DTOs/Course/CreateCourseRequest.cs
@@ -2,6 +2,9 @@
 // Archivo: CreateCourseRequest.cs
 // Proyecto: RH365.Core
 // Ruta: RH365.Core/Application/Features/DTOs/Course/CreateCourseRequest.cs
+// Descripción: DTO para crear un nuevo curso.
+//   - Contiene solo campos editables por el usuario
+//   - Auditoría ISO 27001 se asigna automáticamente
 // ============================================================================
 using System;
 
@@ -12,7 +15,7 @@ namespace RH365.Core.Application.Features.DTOs.Course
         public string CourseCode { get; set; } = null!;
         public string Name { get; set; } = null!;
         public long CourseTypeRefRecID { get; set; }
-        public int? ClassRoomRefRecID { get; set; }
+        public long? ClassRoomRefRecID { get; set; }
         public string? Description { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
